feat: implement create listening and reading test pages, update form handling and validation
