+++ app/test/layout.tsx
@@ -9,14 +9,14 @@ import { MdPreview } from "react-icons/md";
 import { useFormStore } from "@/store/form-store";
 
 const TestLayout = ({ children }: { children: React.ReactNode }) => {
-  const { triggerSubmit } = useFormStore();
+  const { triggerSubmit, title } = useFormStore();
 
   return (
     <div className="mx-auto w-full max-w-screen-2xl px-4 md:px-8 lg:px-14">
       <header className="fixed top-0 right-0 left-0 z-10 mx-auto w-full max-w-screen-2xl">
         <div className="border-primary bg-background flex flex-col items-start gap-4 border-b px-4 pt-12 pb-6 md:px-8 lg:flex-row lg:items-center lg:justify-between lg:px-14">
           <h1 className="text-[clamp(1.5rem,5vw,2.3rem)] font-bold text-white">
-            Create a New Reading Test
+            {title}
           </h1>
 
           <div className="flex items-center gap-3">
