+++ app/test/listening/create/page.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import CreateListeningTestPage from "@/features/test/reading/pages/create-listening-test";
+
+const ListeningTest = () => {
+  return <CreateListeningTestPage />;
+};
+
+export default ListeningTest;
