+++ app/test/reading/create/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import CreateTestPage from "@/features/test/reading/pages/create-test-page";
+import CreateReadingTestPage from "@/features/test/reading/pages/create-reading-test";
 import React from "react";
 
 const ReadingTest = () => {
-  return <CreateTestPage />;
+  return <CreateReadingTestPage />;
 };
 
 export default ReadingTest;
